requirements.txt
aiosqlite==0.18.0
amqp==5.1.1
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
backoff==2.2.1
billiard==3.6.4.0
black==22.12.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.88.0
flower==1.2.0
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.57.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.2
httplib2==0.21.0
httptools==0.5.0
httpx==0.23.1
humanize==4.4.0
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
libsass==0.22.0
Mako==1.2.4
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oauthlib==3.2.2
orjson==3.8.3
pathspec==0.10.3
platformdirs==2.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.7
PyYAML==6.0
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
tomli==2.0.1
tornado==6.2
typing_extensions==4.4.0
ujson==5.6.0
uritemplate==4.1.1
urllib3==1.26.13
uvicorn==0.20.0
vine==5.0.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.4
zipp==3.11.0
